+++ CHANGELOG.md
@@ -8,6 +8,7 @@ _These will be removed in the next major release_
 
 ## 0.27.0
 - Added `Add-BitbucketRepositoryBranch` to create a new branch in a repo
+- Added `Get-BitbucketRepositoryBranch` to return branches in a repo
 - Added `Get-BitbucketRepositoryBranchModel` to get Branching Model in a repo
 - Added `Set-BitbucketRepositoryBranchModel` to modify Branching Model in a repo
 - Added `New-BitbucketRepositoryBranchRestrictionPermissionCheck`